requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Security
bcrypt
